+++ changelog.md
@@ -6,6 +6,7 @@ Version 19.01
 
 + Change the function name of triangle formulas.
 + Remove the external color format of the VPoint color attribute.
++ Remove the "four bar constraint" option of planar linkage class.
 
 Version 18.12
 ===